requirements.txt
Django==5.0.6
psycopg[binary]==3.1.19
python-decouple==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
celery==5.3.6
redis==5.0.4
